requirements.txt
-i https://pypi.org/simple
aiofiles==23.1.0 ; python_version >= '3.7' and python_version < '4.0'
aiosqlite==0.19.0 ; python_version >= '3.7'
asgiref==3.7.2 ; python_version >= '3.7'
asyncua==1.0.2
blinker==1.6.2 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
cffi==1.15.1
channels==4.0.0 ; python_version >= '3.7'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
colorama==0.4.6 ; platform_system == 'Windows'
cryptography==41.0.1 ; python_version >= '3.7'
dash==2.9.3 ; python_version >= '3.6'
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
django==4.2.2
django-background-tasks-updated==1.2.7
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap5==23.3
django-htmx==1.14.0
django-plotly-dash==2.2.1
docker==6.1.3
dpd-components==0.1.0
dpd-static-support==0.0.5
flask==2.3.2 ; python_version >= '3.8'
idna==3.4 ; python_version >= '3.5'
influxdb-client[extra]==1.36.1
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
numpy==1.24.3
packaging==23.1 ; python_version >= '3.7'
pandas==2.0.2
plotly==5.14.1 ; python_version >= '3.6'
pycparser==2.21
pydantic==1.10.8 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3
pywin32==306 ; sys_platform == 'win32'
reactivex==4.0.4 ; python_version >= '3.7' and python_version < '4.0'
requests==2.31.0 ; python_version >= '3.7'
setuptools==67.8.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sortedcontainers==2.4.0
sqlparse==0.4.4 ; python_version >= '3.5'
tenacity==8.2.2 ; python_version >= '3.6'
typing-extensions==4.6.3 ; python_version >= '3.7'
tzdata==2023.3 ; sys_platform == 'win32'
urllib3==2.0.3 ; python_version >= '3.7'
websocket-client==1.5.2 ; python_version >= '3.7'
werkzeug==2.3.4 ; python_version >= '3.8'
whitenoise==6.4.0
